requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
PyMuPDF==1.18.22
python-docx==0.8.11
